Improvements into internal engine time management. Close #131
Also added flag EIF_ENABLE_PER_FRAME_UPDATE.

+++ src/engine/platform/windows/PlatformUtils.cpp
@@ -330,8 +330,8 @@ uint64 GetPerfTimer()
 	//DWORD prev_mask = SetThreadAffinityMask(GetCurrentThread(), 0x0001);
 	QueryPerformanceCounter(&perf_counter);
 	//SetThreadAffinityMask(GetCurrentThread(), prev_mask);
-	
-	return (uint64)(1000000 * perf_counter.QuadPart / perfFreq.QuadPart);
+
+	return (1000000 * perf_counter.QuadPart) / perfFreq.QuadPart;
 }
 
 void OutputDebugTxt(const char *pcTxt)
